Move `impl_binary_value_for_message` macros to proto module. (#1385)
