+++ examples/cryptocurrency/src/lib.rs
@@ -119,8 +119,6 @@ pub mod schema {
 
 /// Transactions.
 pub mod transactions {
-    use std::borrow::Cow;
-
     use super::proto;
     use super::service::SERVICE_ID;
     use exonum::{
