+++ exonum/src/node/mod.rs
@@ -1111,8 +1111,6 @@ impl Node {
 
 #[cfg(test)]
 mod tests {
-    use std::borrow::Cow;
-
     use super::*;
     use crate::blockchain::{
         ExecutionResult, Schema, Service, Transaction, TransactionContext, TransactionSet,
@@ -1121,10 +1119,7 @@ mod tests {
     use crate::events::EventHandler;
     use crate::helpers;
     use crate::proto::{schema::tests::TxSimple, ProtobufConvert};
-    use exonum_merkledb::{
-        impl_binary_value_for_message, BinaryValue, Database, Snapshot, TemporaryDB,
-    };
-    use protobuf::Message as ProtobufMessage;
+    use exonum_merkledb::{BinaryValue, Database, Snapshot, TemporaryDB};
 
     const SERVICE_ID: u16 = 0;
 
@@ -1134,7 +1129,7 @@ mod tests {
         TxSimple(TxSimple),
     }
 
-    impl_binary_value_for_message! { TxSimple }
+    impl_binary_value_for_pb_message! { TxSimple }
 
     impl Transaction for TxSimple {
         fn execute(&self, _: TransactionContext) -> ExecutionResult {
