+++ exonum/src/sandbox/config_updater.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::borrow::Cow;
-
 pub use crate::proto::schema::tests::TxConfig;
 
 use crate::blockchain::{
@@ -24,8 +22,7 @@ use crate::crypto::{Hash, PublicKey, SecretKey};
 use crate::helpers::Height;
 use crate::messages::{Message, RawTransaction, Signed};
 use crate::proto::ProtobufConvert;
-use exonum_merkledb::{impl_binary_value_for_message, BinaryValue, Snapshot};
-use protobuf::Message as PbMessage;
+use exonum_merkledb::{BinaryValue, Snapshot};
 
 pub const CONFIG_SERVICE: u16 = 1;
 
@@ -68,7 +65,7 @@ impl Transaction for TxConfig {
     }
 }
 
-impl_binary_value_for_message! { TxConfig }
+impl_binary_value_for_pb_message! { TxConfig }
 
 impl Service for ConfigUpdateService {
     fn service_name(&self) -> &str {
