+++ exonum/tests/explorer/blockchain/mod.rs
@@ -16,8 +16,6 @@
 
 use futures::sync::mpsc;
 
-use std::borrow::Cow;
-
 use exonum::{
     blockchain::{
         Blockchain, ExecutionError, ExecutionResult, Schema, Service, Transaction,
