+++ testkit/tests/inflating_currency/inflating_cryptocurrency.rs
@@ -11,8 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use std::borrow::Cow;
-
 use exonum_merkledb::{IndexAccess, MapIndex, Snapshot};
 
 use exonum::{
