+++ testkit/tests/service_hooks/hooks.rs
@@ -24,12 +24,9 @@ use exonum::{
 };
 use exonum_merkledb::Snapshot;
 
-use std::{
-    borrow::Cow,
-    sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc,
-    },
+use std::sync::{
+    atomic::{AtomicUsize, Ordering},
+    Arc,
 };
 
 pub const SERVICE_ID: u16 = 512;
